+++ src/insitu-shuffle.c
@@ -75,7 +75,7 @@ SEXP insitu_shuffle_fast_(SEXP x_)  {
     int n = length(x_);
     int *x = INTEGER(x_);
     for (int i = n-1; i > 0; i--)  {
-      int j = (int)random_integer_on_interval(i + 1);
+      int j = (int)random_integer_on_interval((uint64_t)i + 1);
       int tmp = x[i];
       x[i] = x[j];
       x[j] = tmp;
@@ -85,7 +85,7 @@ SEXP insitu_shuffle_fast_(SEXP x_)  {
     int n = length(x_);
     double *x = REAL(x_);
     for (int i = n-1; i > 0; i--)  {
-      int j = (int)random_integer_on_interval(i + 1);
+      int j = (int)random_integer_on_interval((uint64_t)i + 1);
       double tmp = x[i];
       x[i] = x[j];
       x[j] = tmp;
@@ -94,7 +94,7 @@ SEXP insitu_shuffle_fast_(SEXP x_)  {
   case STRSXP: {
     int n = length(x_);
     for (int i = n-1; i > 0; i--)  {
-      int j = (int)random_integer_on_interval(i + 1);
+      int j = (int)random_integer_on_interval((uint64_t)i + 1);
       SEXP tmp = STRING_ELT(x_, i);
       SET_STRING_ELT(x_, i, STRING_ELT(x_, j));
       SET_STRING_ELT(x_, j, tmp);
