+++ src/insitu-sort.c
@@ -55,10 +55,10 @@ SEXP insitu_sort_(SEXP x_, SEXP decreasing_) {
 
   switch(TYPEOF(x_)) {
   case INTSXP: {
-    qsort(INTEGER(x_), length(x_), sizeof(int), asLogical(decreasing_) ? int_cmpfunc_dec : int_cmpfunc_inc);
+    qsort(INTEGER(x_), (size_t)length(x_), sizeof(int), asLogical(decreasing_) ? int_cmpfunc_dec : int_cmpfunc_inc);
   } break;
   case REALSXP: {
-    qsort(REAL(x_), length(x_), sizeof(double), asLogical(decreasing_) ? dbl_cmpfunc_dec : dbl_cmpfunc_inc);
+    qsort(REAL(x_), (size_t)length(x_), sizeof(double), asLogical(decreasing_) ? dbl_cmpfunc_dec : dbl_cmpfunc_inc);
   } break;
   default:
     error("insitu_sort(): type not supported");
